- nested style for  studio
- Added StudioContainer
