+++ src/shared/App.less
@@ -59,10 +59,8 @@
   display: flex;
   margin-left: 50px;
   width: 100%;
-}
-
-.studio-view {
   .workspace {
     display: flex;
   }
 }
+
