+++ src/shared/views/StudioView.tsx
@@ -3,48 +3,25 @@ import { match } from 'react-router';
 import { useAsyncEffect } from 'use-async-effect';
 import { Resource } from '@bbp/nexus-sdk';
 import { useNexusContext } from '@bbp/react-nexus';
-import WorkspaceList from '../containers/WorkspaceList';
+import StudioContainer from '../containers/StudioContainer';
 
 type StudioViewProps = {
   match: match<{ orgLabel: string; projectLabel: string; studioId: string }>;
-}
+};
 
 const StudioView: React.FunctionComponent<StudioViewProps> = props => {
   const { match } = props;
-  const [isStudio, setIsStudio] = React.useState<boolean>(false);
-  const [workSpaceIds, setWorkspaceIds] = React.useState<string[]>([]);
   const nexus = useNexusContext();
   const {
     params: { orgLabel, projectLabel, studioId },
   } = match;
 
-  useAsyncEffect(async () => {
-    const activeResource = (await nexus.Resource.get(
-      orgLabel,
-      projectLabel,
-      studioId
-    )) as Resource;
-    if (activeResource['@type'] === 'Studio') {
-      setIsStudio(true);
-    }
-    const workspaceIds: string[] = activeResource['workspaces'];
-    setWorkspaceIds(workspaceIds);
-  }, [studioId]);
   return (
-    <h4>
-      {' '}
-      {isStudio ? (
-        <div className="studio-view">
-          <WorkspaceList
-            orgLabel={orgLabel}
-            projectLabel={projectLabel}
-            workSpaceIds={workSpaceIds}
-          />{' '}
-        </div>
-      ) : (
-        'The Resource is not a Studio'
-      )}
-    </h4>
+    <StudioContainer
+      orgLabel={orgLabel}
+      projectLabel={projectLabel}
+      studioId={studioId}
+    />
   );
 };
